Improvements to localization
OpenWeather has two descriptions of the current weather: the "main" and the "description". In English, the "main" value looks better on a widget. But it is not translated, even when a language is provided - the description must be used instead.

Additionally, OpenWeather supports far fewer languages than iOS. This code attempts to find the closest possible language match by looking at the hard-coded locale, the device's locale, and both versions without any suffixes added. The rationale here is that getting close is better than not translating at all.