Updates from @suxus-zz
Thanks to @suxus-zz, this update allows for a new COVID data element, a rain percentage in the future weather view, and the Weather.com links to be in the correct locale.
I also fixed an issue where updating the code broke if you didn't have the "minutesAfter" setting set. And, I alphabetized the function list.