Fixed dateDiff
The StackOverflow for this is a mess. My method was broken because all-day events are at midnight, and for whatever reason rounding it went to 0 (meaning same day). This is so easily fixed by just coercing all of the dates to midnight - so even with hour/DST weirdness, it just works.