Add COVID token for yesterday’s cases and deaths
I retrieve daily historical COVID numbers from a different API and combine a few of those numbers with the COVID data from the original API. Note that I rename the keys from the alternate API to match the conventions in the original API.

I also added the two new COVID tokens (`{yesterdayCases}` and `{yesterdayDeaths}`) to the pop-up description when editing the COVID text.

TODO: Make the country for the alternate API a variable. It’s currently set to “us”, and unfortunately the country names in the alternate API don’t necessarily match those in the original API, so a separate country option would be required, or we’d need to map countries between the two.